+++ bishop.py
@@ -1,0 +1,8 @@
+#se mueve solo en diagonales en cualquier casilla
+from piece import Piece
+
+class Bishop(Piece):
+   
+   def mover(self):
+      pass
+