+++ chess/king.py
@@ -1,0 +1,9 @@
+#se mueve vertical, horizontal y vertical pero solo de a una casilla
+from chess.piece import Piece
+
+class King(Piece):
+   white_str = "♔" 
+   black_str = "♚"
+   
+ 
+