+++ chess/knight.py
@@ -1,0 +1,9 @@
+#se mueve en forma de L y puede saltearse otras piezas
+from chess.piece import Piece
+
+class Knight(Piece): 
+   white_str = "♘"  
+   black_str = "♞"
+   
+   
+