+++ chess/queen.py
@@ -1,0 +1,9 @@
+#se mueve vertical,horizontal y diagonal en cualquier casilla
+from chess.piece import Piece
+
+class Queen(Piece):
+   white_str = "♕" 
+   black_str = "♛"
+    
+  
+